Fixed cmake, see below
First changed the command to generate the resource file to work with
out of tree builds.

Reformatted some stuff.

Set the C++ standard for the executable to be c++11. This makes it so
that it builds on my system where the compiler is set to clang++.

Changed to install to the relative bin directory so that the user can
specify CMAKE_INSTALL_PREFIX. THIS WILL MEAN PKGBUILDS WILL HAVE TO
CHANGE TO SPECIFY CMAKE_INSTALL_PREFIX.