Made a few improvements, see below
Updated CMakeLists to be consistent with spaces before open parenthesis.

Changed default font to Serif because that's what 10FastFingers does
and it looks better when the font is so big.

Updated window names, but that might have been the last commit.

Added a basic desktop icon so that it can be launched from desktop
environments. Didn't work up until now because all developers were
using dmenu with i3.

Added a default window size of 1280x720 for non-tiling window managers.