Made a bunch of changes, see below
Removed a debug message that might have been there before.

Set the history title dialog and transient for.

Cleaned up the calculateScore method by moving parts of it into two
functions which update the trouble words file and history file
respectively.

Added a lock for the trouble words file and updated it to use the swap
mechanism.