Various things, see commit body.
Fixed bug on GCC where it would segfault. One thing I realized was that
because I was using a hash object I didn't need to provide an argument
to the constructor for two classes anymore and I was able to let the
number in the unordered collections constructors go to their defaults.
However, the segfault stopped occuring when I removed a call to
TextBuffer{} in the initializers for those classes.

Also changed the text buffer system to fix some issues with stickers.
There is now a variable associated with the buffer for adding newlines
which is meant to be sent by the caller when they insert text for the
correct behavior. I also made a bunch of tweaks to the replacement code
so that I'm 70% sure it works now.